Share decomposition between NVIDIA and AMD paths
