+++ include/triton/Dialect/TritonGPU/Transforms/DecomposeScaledBlocked.h
@@ -1,0 +1,8 @@
+#include "mlir/IR/PatternMatch.h"
+
+namespace mlir::triton::gpu {
+
+void populateDecomposeScaledBlockedPatterns(mlir::RewritePatternSet &patterns,
+                                            int benefit);
+
+} // namespace mlir::triton::gpu
